Add method hasBag to ViewErrorBag
When redirecting withErrors($validator) the edit form can show individual messages. But to show a generic message, something like: please check the red marked field below I don't know anything else but to flash a extra variable. This method would remedy that. Now I could check: $errors->hasBag() to see if the form is a redirect from a post with a errors message bag or a first get for input.