Do case insensitive search using str::contains
Initially you can only use Str::contains but it only returns true if the matching words or characters are in the same case.
I added an optional $case_sensitive variable that is by default true. If passed as false the search will be case_insensitive.
When false, it only converts the needle and haystack to lowercase and continues with the search.